api/requirements.txt
anyio==3.6.2
bson==0.5.10
click==8.1.3
colorama==0.4.6
decouple==0.0.7
dnspython==2.3.0
fastapi==0.95.1
h11==0.14.0
idna==3.4
motor==3.1.2
pydantic==1.10.7
PyJWT==2.7.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
